+++ Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 let package = Package(
     name: "SrizonVoice",
     platforms: [
-        .macOS(.v13)
+        .macOS(.v12)
     ],
     products: [
         .executable(name: "SrizonVoice", targets: ["SrizonVoice"])